+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- `Config` doesn't force OpenGL `Api` by default.
+- `Display::create_context` now uses the most recent available `Api` from the `Config` when `ContextApi` is not specified in `ContextAttributes`.
+
 # Version 0.30.0-beta.2 (2022-09-03)
 
 - macOS: Fix memory leak in `Surface`.